A basic Hash Map program
Putting values into a hash map from a text file where the first value of the file is key and other three values are grouped together into an object and insert into the hash map. Task 1 is to display the element of the hash map. Task 2 is to take a randomly generated integer value for a certain range and compare with the hash map key value. If the value matches, delete that value from the list and display the updated hash map values.